+++ js/views/index.js
@@ -1,0 +1,9 @@
+/**
+ * WOW3 Views Index
+ * Central export point for all view classes
+ */
+
+export { UIManager } from './UIManager.js';
+export { RightSidebar } from './RightSidebar.js';
+export { StatusBar } from './StatusBar.js';
+export { ElementsTree } from './ElementsTree.js';